pcap_findalldevs() can fail on windows 7
The first time in a process that pcap_findalldevs() is called on a
windows system, it frequently fails with

  PacketGetAdapterNames: The data area passed to a system call is too small. (122)

Subsequent calls succeed.  In Cygwin builds, wrap this function with a
loop that tries once per second and doesn't exit the argus process until
a threshold has been reached (10 in this case).