ArgusSourceProcess: close pcap source when ArgusGetPackets exits
Whenever a source thread (ArgusGetPackets) exits, close the pcap
source it was using and remove that source structure from the
stask->srcs[] array.  Adjust thread and source counters as necessary.
Clear the model's output list pointer to prevent the output thread
from being stopped.

A source thread will exit when an interface is configured down or
removed.  ArgusSourceProcess will spawn a new source thread if the
interface is configured back up or another interface is added (and
the "all" interface was specified on the commandline).

Trac ticket #37