choose abbreviated interface name for srcid
Add file ArgusIfnam.c with logic to choose a unique four-character
interface name abbreviation to use in the source ID.  If an
abbreviation has already been used, advance the first letter to the
next in the alphabet.  For example, if "e123" has been used, the
next interface name that is abbreviated to the same string will be
replaced with "f123".  Windows "transport names" are handled by
skipping over the leading backslash and truncating the name to
remove the trailing brace.

Trac ticket #33