initial FlexLM licensing support for Argus
Adds argus_lic.[ch].  The flexnet publisher distribution's location
is passed to the configure script with the --with-flexlm= parameter.
This should be the top level of the flexlm directory tree.  The
FlexLM directory should be writeable because it may be necessary
to create the lm_new.o object file and this is handled by the
makefile that Flexera provides.

Running the pre-built flexlm binaries requires the redhat-lsb-core
RPM to be installed.