argus@.service systemd template for NUMA support
This service allows an instance of argus to be bound to a NUMA node
and to only monitor Napatech queues allocated on that node.  Starting
argus@0 will launch an argus process bound to numa node zero.  The
selection of napatech "device" names is not very automatic.  It's
assumed that contiguous ranges of device numbers (nt3g0, nt3g1,
nt3g2, etc.) are allocated on the same node and that every numa
node has the same number of napatech queues.

Starting argus@all will spawn an argus process that has no numa
bindings and deals only with management interfaces and non-napatech
(or non-hardware-accelerated) capture interfaces.

These services are compatible with argus.service.