make argus-setup-numa work for management interfaces
The syntax for identifying numa nodes in the "numactl" program includes
the string "all" to indicate that no binding to nodes should occur.  Use
this syntax with the argus@.service template to indicate that argus should
be started for interfaces that do not need any particular node binding.
These are management interfaces and non-accelerated monitoring interfaces.