fix loop bounds in ArgusSourceProcess
There is a case where a loop's iteration count is limited by a
variable named "count" and the value of "count" is clobbered inside
the loop body.  Provide a separate variable for the needs of the
loop body.  Change a subsequent loop to use the same variable name
for stylistic consistency.