use modern format in distribution tar file
The -o option tells (gnu) tar to create v7 output.  This has a max
filename length of 99 characters.  Tar fails because of the filename
lengths in the pkg/osx directory.  BSD tar generates POSIX/ustar output
with that option, so the behavior is not consistent across platforms.

GNU tar is portable; allow long filenames.