ignore directory components in pcap filename for device string
When argus is given a pcap file for input, the source task's device name
(ArgusDeviceStr) is the supplied filename.  If another source is later
discovered, any forward slashes in the source task's name are confused
with the slash that separates the srcid and interface name, if provided
(i.e., "-i interface[/srcid]").

Remove directory components from filename for use as a device string.

Prevents errors such as:
     ArgusInfo: 02 Feb 18 11:00:20.200307 mapping interface name /tmp/spillway-all.cap -> tcap
     ArgusInfo: 02 Feb 18 11:00:22.656832 ArgusSourceProcess: new device: bluetooth0 found
    ArgusError: 02 Feb 18 11:00:22.656850 ArgusParseSourceID error: tmp/spillway-all.cap: interface len gt 4 char

Also move the copy of parameter cmd in setArgusDevice() to somewhere
after the check that cmd is not NULL.