work around gcc optimizer confusion
ArgusGenerateRecord() confuses the gcc 7 optimizer, which re-orders
instructions, and the wire-format packet does not always have the most
up-to-date contents.  "Touch" the problematic memory location to force
the compiler to order the instructions so that we get correct flow
records on the wire.

The problem can be circumvented by setting
ARGUS_GENERATE_BIDIRECTIONAL_TIMESTAMPS=yes in argus.conf so that the
code path in question is avoided.

compiled with options "-mtune=generic -march=x86-64 -g -O3 -O2
-fexceptions -fstack-protector-strong --param ssp-buffer-size=4" using
gcc (GCC) 7.2.1 20170915 (Red Hat 7.2.1-2)