do not temporarily remove devices from ArgusSourceStruct lists
Multiple threads access the ArgusSourceStruct's ArgusDeviceList.  If one
thread has removed an element from the device list with the intention
of putting it back later (even with ARGUS_LOCK), it is possible for the
ArgusSourceProcess() thread to dereference a null pointer while comparing
the known devices to a list obtained from libpcap.

The sequence of events is something similar to
Thread 1:
  ArgusGetPackets()
     ArgusGetInterfaceStatus()
        device = ArgusPopFrontList(src->ArgusDeviceList, ARGUS_LOCK) # removes the only element

Thread 2:
  ArgusSourceProcess()
     device = ArgusPopFrontList(src->ArgusDeviceList, ARGUS_LOCK) # returns NULL

Thread 1:
        ArgusPushFrontList(src->ArgusDeviceList, device, ARGUS_LOCK)

Thread 2:
      strcmp(device->name, ...) # crash