return copy of device name in getArgusDevice
The device name is part of ArgusDeviceStruct and protected by
ArgusDeviceList->lock.  The caller cannot hold the this lock since this
function aquires it when retrieving the device name.  To avoid having
a pointer to the name without the protection of the list lock, make a
copy of the name and return that, instead.

(cherry picked from commit 1608a1ab7cc6ba636126141e97e594ea158d530a)